Moves memrev function to common.
